Misc cleanup commit
