+++ inst/abstract-snake-tree/src/NodeTypes/Parsed_Ui_Node.ts
@@ -7,7 +7,6 @@ type Node_Fields_By_Type = {
   call: {
     fn_name: string;
     args: Parsed_Ui_Node[];
-    known: boolean;
   };
   string: {
     value: string;
