+++ inst/abstract-snake-tree/src/python_ast_to_shiny_ui_node.test.ts
@@ -50,39 +50,6 @@ export const basicNavbarPage = {
   ],
 } satisfies KnownShinyUiNode;
 
-const navbar_page_app = `
-from shiny import *
-
-app_ui = ui.page_navbar(
-  ui.nav(
-    "Settings",
-    ui.input_slider(
-      id = "inputId",
-      label = "Slider Input",
-      min = 0,
-      max = 10,
-      value = 5.5,
-      width = "100%"
-    )
-  ),
-  ui.nav(
-    "Plot 1",
-    ui.output_plot(
-      id = "MyPlot",
-      width = "100%",
-      height = "100%"
-    )
-  ),
-  title = "My Navbar Page",
-  collapsible = False
-)
-
-def server(input, output, session):
-  pass
-
-app = App(app_ui, server)
-`;
-
 describe("Can go from parsed-tree-sitter ast to proper ShinyUiNode", () => {
   test("Simple slider input with all named args", () => {
     const sliderInputDef = `
